ASI, HI, TC, ALT working stable

+++ MF_SixPack/MF_ALT/MF_ALT.h
@@ -48,5 +48,5 @@ private:
     void setAltitude(float value);    // angle for heading
     void setBaro(float value); // angle for heading bug
     void setPowerSave(bool enabed);
-    void setInstrumentBrightness(uint8_t value);
+    void setInstrumentBrightness(float value);
 };
